requirements.txt
pytest
requests
requests-mock
python-dotenv
faker
